[REFACTOR] OptionalException 리팩토링
